Consolidate utils.pyfile into utils.openpy
